tools/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=4.5=1_gnu
addict=2.4.0=pypi_0
aiohttp=3.8.1=pypi_0
aiosignal=1.2.0=pypi_0
anykeystore=0.2=pypi_0
apex=0.1=pypi_0
apptools=5.1.0=pypi_0
argon2-cffi=21.1.0=pypi_0
async-generator=1.10=pypi_0
async-timeout=4.0.1=pypi_0
asynctest=0.13.0=pypi_0
attrs=21.2.0=pypi_0
backcall=0.2.0=pypi_0
black=21.11b1=pypi_0
blas=1.0=mkl
bleach=4.1.0=pypi_0
ca-certificates=2021.10.26=h06a4308_2
cachetools=4.2.4=pypi_0
certifi=2021.5.30=py36h06a4308_0
cffi=1.15.0=pypi_0
charset-normalizer=2.0.8=pypi_0
click=8.0.3=pypi_0
configobj=5.0.6=pypi_0
cryptacular=1.6.2=pypi_0
cudatoolkit=10.2.89=hfd86e86_1
cycler=0.11.0=pypi_0
cython=0.29.24=pypi_0
dataclasses=0.8=pypi_0
decorator=4.4.2=pypi_0
defusedxml=0.7.1=pypi_0
descartes=1.1.0=pypi_0
det3d=1.0rc0+7a45493=dev_0
easydict=1.9=pypi_0
entrypoints=0.3=pypi_0
envisage=6.0.1=pypi_0
fire=0.4.0=pypi_0
flake8=4.0.1=pypi_0
freetype=2.11.0=h70c0345_0
frozenlist=1.2.0=pypi_0
greenlet=1.1.2=pypi_0
hupper=1.10.3=pypi_0
idna=3.3=pypi_0
idna-ssl=1.1.0=pypi_0
imagecorruptions=1.1.2=pypi_0
imageio=2.12.0=pypi_0
importlib-metadata=4.2.0=pypi_0
importlib-resources=5.4.0=pypi_0
iniconfig=1.1.1=pypi_0
intel-openmp=2021.4.0=h06a4308_3561
iou3d=0.0.0=pypi_0
ipdb=0.13.9=pypi_0
ipykernel=5.5.6=pypi_0
ipython=7.16.2=pypi_0
ipython-genutils=0.2.0=pypi_0
ipywidgets=7.6.5=pypi_0
jedi=0.17.2=pypi_0
jinja2=3.0.3=pypi_0
joblib=1.1.0=pypi_0
jpeg=9d=h7f8727e_0
jsonschema=3.2.0=pypi_0
jupyter=1.0.0=pypi_0
jupyter-client=7.1.0=pypi_0
jupyter-console=6.4.0=pypi_0
jupyter-core=4.9.1=pypi_0
jupyterlab-pygments=0.1.2=pypi_0
jupyterlab-widgets=1.0.2=pypi_0
kiwisolver=1.3.1=pypi_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.35.1=h7274673_9
libffi=3.3=he6710b0_2
libgcc-ng=9.3.0=h5101ec6_17
libgomp=9.3.0=h5101ec6_17
libpng=1.6.37=hbc83047_0
libstdcxx-ng=9.3.0=hd4cf53a_17
libtiff=4.2.0=h85742a9_0
libwebp-base=1.2.0=h27cfd23_0
lltm-cpp=0.0.0=pypi_0
llvmlite=0.36.0=pypi_0
lyft-dataset-sdk=0.0.8=pypi_0
lz4-c=1.9.3=h295c915_1
markupsafe=2.0.1=pypi_0
matplotlib=3.3.4=pypi_0
mayavi=4.7.4=pypi_0
mccabe=0.6.1=pypi_0
mistune=0.8.4=pypi_0
mkl=2020.2=256
mkl-service=2.3.0=py36he8ac12f_0
mkl_fft=1.3.0=py36h54f3939_0
mkl_random=1.1.1=py36h0573a6f_0
multidict=5.2.0=pypi_0
mypy-extensions=0.4.3=pypi_0
nbclient=0.5.9=pypi_0
nbconvert=6.0.7=pypi_0
nbformat=5.1.3=pypi_0
ncurses=6.3=h7f8727e_2
nest-asyncio=1.5.1=pypi_0
networkx=2.5.1=pypi_0
ninja=1.10.2=h5e70eb0_2
notebook=6.4.6=pypi_0
numba=0.53.1=pypi_0
numpy=1.19.2=py36h54aff64_0
numpy-base=1.19.2=py36hfa32c7d_0
nuscenes-devkit=1.1.9=pypi_0
oauthlib=3.1.1=pypi_0
objgraph=3.5.0=pypi_0
olefile=0.46=py36_0
open3d-python=0.7.0.0=pypi_0
opencv-contrib-python=4.5.4.60=pypi_0
opencv-python=4.5.4.60=pypi_0
openjpeg=2.4.0=h3ad879b_0
openssl=1.1.1l=h7f8727e_0
packaging=21.3=pypi_0
pandas=1.1.5=pypi_0
pandocfilters=1.5.0=pypi_0
parso=0.7.1=pypi_0
pastedeploy=2.1.1=pypi_0
pathspec=0.9.0=pypi_0
pbkdf2=1.3=pypi_0
pexpect=4.8.0=pypi_0
pickleshare=0.7.5=pypi_0
pillow=8.4.0=pypi_0
pip=21.2.2=py36h06a4308_0
plaster=1.0=pypi_0
plaster-pastedeploy=0.7=pypi_0
platformdirs=2.4.0=pypi_0
plotly=5.4.0=pypi_0
pluggy=1.0.0=pypi_0
prometheus-client=0.12.0=pypi_0
prompt-toolkit=3.0.23=pypi_0
protobuf=3.19.1=pypi_0
ptyprocess=0.7.0=pypi_0
py=1.11.0=pypi_0
pyaml=21.10.1=pypi_0
pybind11=2.8.1=pypi_0
pycocotools=2.0.3=pypi_0
pycodestyle=2.8.0=pypi_0
pycparser=2.21=pypi_0
pyface=7.4.1=pypi_0
pyflakes=2.4.0=pypi_0
pygments=2.10.0=pypi_0
pyparsing=3.0.6=pypi_0
pyqt5=5.15.6=pypi_0
pyqt5-qt5=5.15.2=pypi_0
pyqt5-sip=12.9.1=pypi_0
pyquaternion=0.9.9=pypi_0
pyramid=2.0=pypi_0
pyramid-mailer=0.15.1=pypi_0
pyrsistent=0.18.0=pypi_0
pytest=6.2.5=pypi_0
pytest-runner=5.3.1=pypi_0
python=3.6.13=h12debd9_1
python-dateutil=2.8.2=pypi_0
python-graphviz=0.18.2=pypi_0
python3-openid=3.2.0=pypi_0
pytorch=1.5.0=py3.6_cuda10.2.89_cudnn7.6.5_0
pytz=2021.3=pypi_0
pywavelets=1.1.1=pypi_0
pyyaml=6.0=pypi_0
pyzmq=22.3.0=pypi_0
qtconsole=5.2.1=pypi_0
qtpy=1.11.2=pypi_0
readline=8.1=h27cfd23_0
regex=2021.11.10=pypi_0
repoze-sendmail=4.4.1=pypi_0
requests=2.26.0=pypi_0
requests-oauthlib=1.3.0=pypi_0
scikit-image=0.17.2=pypi_0
scikit-learn=0.24.2=pypi_0
scipy=1.5.4=pypi_0
send2trash=1.8.0=pypi_0
setuptools=58.0.4=py36h06a4308_0
shapely=1.8.0=pypi_0
six=1.16.0=pyhd3eb1b0_0
spconv=1.0=pypi_0
sqlalchemy=1.4.27=pypi_0
sqlite=3.36.0=hc218d9a_0
tenacity=8.0.1=pypi_0
tensorboardx=2.4.1=pypi_0
termcolor=1.1.0=pypi_0
terminado=0.12.1=pypi_0
terminaltables=3.1.0=pypi_0
testpath=0.5.0=pypi_0
threadpoolctl=3.0.0=pypi_0
tifffile=2020.9.3=pypi_0
tk=8.6.11=h1ccaba5_0
toml=0.10.2=pypi_0
tomli=1.2.2=pypi_0
torchvision=0.6.0=py36_cu102
tornado=6.1=pypi_0
tqdm=4.62.3=pypi_0
traitlets=4.3.3=pypi_0
traits=6.3.2=pypi_0
traitsui=7.3.1=pypi_0
transaction=3.0.1=pypi_0
translationstring=1.4=pypi_0
typed-ast=1.5.0=pypi_0
typing-extensions=4.0.0=pypi_0
urllib3=1.26.7=pypi_0
velruse=1.1.1=pypi_0
venusian=3.0.0=pypi_0
vtk=8.1.2=pypi_0
wcwidth=0.2.5=pypi_0
webencodings=0.5.1=pypi_0
webob=1.8.7=pypi_0
wheel=0.37.0=pyhd3eb1b0_1
widgetsnbextension=3.5.2=pypi_0
wslink=1.1.1=pypi_0
wtforms=3.0.0=pypi_0
wtforms-recaptcha=0.3.2=pypi_0
xz=5.2.5=h7b6447c_0
yarl=1.7.2=pypi_0
zipp=3.6.0=pypi_0
zlib=1.2.11=h7b6447c_3
zope-deprecation=4.4.0=pypi_0
zope-interface=5.4.0=pypi_0
zope-sqlalchemy=1.6=pypi_0
zstd=1.4.9=haebb681_0
